[JENKINS-54459] Rename constants and its visibility
Rename the constants to do the retries and change their visibilities.

Change the package of the test to be able to access the constants in
order to redefine their values in the tests.
